setup availability demo routing and rendering
